+++ src/xray/xr_3da/xrGame/Physics.cpp
@@ -133,8 +133,8 @@ IC static int CollideIntoGroup(dGeomID o1, dGeomID o2,dJointGroupID jointGroup,C
 		bool do_collide	=	true;
 		dxGeomUserData* usr_data_1		=NULL;
 		dxGeomUserData* usr_data_2		=NULL;
-		u16				material_idx_1	=0;
-		u16				material_idx_2	=0;
+		u32				material_idx_1	=0;
+		u32				material_idx_2	=0;
 
 		surface.mu =1.f;// 5000.f;
 		surface.soft_erp=1.f;//ERP(world_spring,world_damping);
@@ -144,13 +144,13 @@ IC static int CollideIntoGroup(dGeomID o1, dGeomID o2,dJointGroupID jointGroup,C
 		usr_data_1 = retrieveGeomUserData(g1);
 		usr_data_2 = retrieveGeomUserData(g2);
 		///////////////////////////////////////////////////////////////////////////////////////////////////
-		if(usr_data_2)	material_idx_2=usr_data_2->material;
-		if(usr_data_1)	material_idx_1=usr_data_1->material;
+		if(usr_data_2)	material_idx_2= (u32)usr_data_2->material;
+		if(usr_data_1)	material_idx_1= (u32)usr_data_1->material;
 		bool is_tri_1=dTriListClass == dGeomGetClass(g1);
 		bool is_tri_2=dTriListClass == dGeomGetClass(g2);
 		if(!is_tri_2&&!is_tri_1) surface.mode=0;
-		if(is_tri_1) material_idx_1=(u16)surface.mode;
-		if(is_tri_2) material_idx_2=(u16)surface.mode;
+		if(is_tri_1) material_idx_1=(u32)surface.mode;
+		if(is_tri_2) material_idx_2=(u32)surface.mode;
 		SGameMtl* material_1=GMLib.GetMaterialByIdx(material_idx_1);
 		SGameMtl* material_2=GMLib.GetMaterialByIdx(material_idx_2);
 		////////////////params can be changed in callbacks//////////////////////////////////////////////////////////////////////////
